Rename galaxy.util.logging to galaxy.util.custom_logging
Otherwise there may be import issues:
```
==================================== ERRORS ====================================

________________________ ERROR collecting test session _________________________

.tox/py27-unit/lib/python2.7/site-packages/py/_path/common.py:383: in visit

    for x in Visitor(fil, rec, ignore, bf, sort).gen(self):

.tox/py27-unit/lib/python2.7/site-packages/py/_path/common.py:425: in gen

    if p.check(dir=1) and (rec is None or rec(p))])

.tox/py27-unit/lib/python2.7/site-packages/_pytest/main.py:667: in _recurse

    ihook = self.gethookproxy(dirpath)

.tox/py27-unit/lib/python2.7/site-packages/_pytest/main.py:482: in gethookproxy

    my_conftestmodules = pm._getconftestmodules(fspath)

.tox/py27-unit/lib/python2.7/site-packages/_pytest/config/__init__.py:424: in _getconftestmodules

    mod = self._importconftest(conftestpath.realpath())

.tox/py27-unit/lib/python2.7/site-packages/_pytest/config/__init__.py:463: in _importconftest

    raise ConftestImportFailure(conftestpath, sys.exc_info())

E   ConftestImportFailure: (local('/home/circleci/repo/test/unit/queue_worker/conftest.py'), (<type 'exceptions.AttributeError'>, AttributeError("'module' object has no attribute 'getLogger'",), <traceback object at 0x7fabd96b7eb0>))

----- generated html file: file:///home/circleci/repo/run_unit_tests.html ------

!!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!!

=========================== 1 error in 1.83 seconds ============================

```
