+++ doc/source/lib/galaxy.util.logging.rst
@@ -1,7 +1,7 @@
-galaxy.util.logging package
+galaxy.util.custom_logging package
 ===========================
 
-.. automodule:: galaxy.util.logging
+.. automodule:: galaxy.util.custom_logging
     :members:
     :undoc-members:
     :show-inheritance:
@@ -9,10 +9,10 @@ galaxy.util.logging package
 Submodules
 ----------
 
-galaxy.util.logging.fluent\_log module
+galaxy.util.custom_logging.fluent\_log module
 --------------------------------------
 
-.. automodule:: galaxy.util.logging.fluent_log
+.. automodule:: galaxy.util.custom_logging.fluent_log
     :members:
     :undoc-members:
     :show-inheritance:
