+++ lib/galaxy/jobs/actions/post.py
@@ -11,7 +11,7 @@ from galaxy.util import (
     send_mail,
     unicodify,
 )
-from galaxy.util.logging import get_logger
+from galaxy.util.custom_logging import get_logger
 
 log = get_logger(__name__)
