+++ lib/galaxy/jobs/runners/slurm.py
@@ -11,7 +11,7 @@ import time
 from galaxy import model
 from galaxy.jobs.runners.drmaa import DRMAAJobRunner
 from galaxy.util import unicodify
-from galaxy.util.logging import get_logger
+from galaxy.util.custom_logging import get_logger
 
 log = get_logger(__name__)
