+++ lib/galaxy/util/__init__.py
@@ -65,8 +65,8 @@ try:
 except ImportError:
     uwsgi = None
 
+from .custom_logging import get_logger
 from .inflection import English, Inflector
-from .logging import get_logger
 from .path import safe_contains, safe_makedirs, safe_relpath  # noqa: F401
 
 inflector = Inflector(English)
